src/main/java/com/kucuktech/lab/recipes/request/SearchRequest.java
package com.kucuktech.lab.recipes.request;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.io.Serializable;
import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
@JsonIgnoreProperties(ignoreUnknown = true)
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class SearchRequest implements Serializable {
    private static final long serialVersionUID = -5299079227305887889L;

    private List<FilterRequest> filters;

    public List<FilterRequest> getFilters() {
        return Optional.ofNullable(this.filters)
                .orElse(new ArrayList<>());
    }
}